Comment table/view/schema.

Comment on a table, a view or a schema.

dbComment(
  conn,
  name,
  comment,
  type = c("table", "view", "schema"),
  display = TRUE,
  exec = TRUE
)

Arguments

conn

A connection object.

name

A character string specifying a PostgreSQL table, view or schema name.

comment

A character string specifying the comment.

type

The type of the object to comment, either "table", "view", or "schema"

display

Logical. Whether to display the query (defaults to TRUE).

exec

Logical. Whether to execute the query (defaults to TRUE).

Value

If exec = TRUE, returns TRUE if the comment was successfully applied.

See also

Author

Mathieu Basille basille@ufl.edu

Examples

## examples use a dummy connection from DBI package
conn <- DBI::ANSI()
dbComment(conn, name = c("schema", "table"), comment = "Comment on a view.",
    type = "view", exec = FALSE)
#> Query not executed:
#> COMMENT ON VIEW "schema"."table" IS 'Comment on a view.';
dbComment(conn, name = "test_schema", comment = "Comment on a schema.", type = "schema",
    exec = FALSE)
#> Query not executed:
#> COMMENT ON SCHEMA "test_schema" IS 'Comment on a schema.';