Check and create PostGIS extension.

The function checks for the availability of the PostGIS extension, and if it is available, but not installed, install it. Additionally, can also install Topology, Tiger Geocoder, SFCGAL and Raster extensions.

pgPostGIS(
  conn,
  topology = FALSE,
  tiger = FALSE,
  sfcgal = FALSE,
  raster = FALSE,
  display = TRUE,
  exec = TRUE
)

Arguments

conn

A connection object (required, even if exec = FALSE).

topology

Logical. Whether to check/install the Topology extension.

tiger

Logical. Whether to check/install the Tiger Geocoder extension. Will also install extensions "fuzzystrmatch", "address_standardizer", and "address_standardizer_data_us" if all are available.

sfcgal

Logical. Whether to check/install the SFCGAL extension.

raster

Logical. Whether to check/install the Raster extension

display

Logical. Whether to display the query (defaults to TRUE).

exec

Logical. Whether to execute the query (defaults to TRUE).

Value

If exec = TRUE, returns TRUE if PostGIS is installed.

Author

Mathieu Basille basille@ufl.edu and Adrián Cidre González adrian.cidre@gmail.com

Examples

## 'exec = FALSE' does not install any extension, but nevertheless
## check for available and installed extensions:
if (FALSE) {
    pgPostGIS(con, topology = TRUE, tiger = TRUE, sfcgal = TRUE,
        exec = FALSE)
}