R interface to a PostGIS database.

'rpostgis' provides an interface between R and 'PostGIS'-enabled 'PostgreSQL' databases to transparently transfer spatial data. Both vector (points, lines, polygons) and raster data are supported in read and write modes. Also provides convenience functions to execute common procedures in 'PostgreSQL/PostGIS'. For a list of documented functions, use library(help = "rpostgis").

Details

A typical session starts by establishing the connection to a working PostgreSQL database:

library(rpostgis) con <- dbConnect("PostgreSQL", dbname = <dbname>, host = <host>, user = <user>, password = <password>)

For example, this could be:

con <- dbConnect("PostgreSQL", dbname = "rpostgis", host = "localhost", user = "postgres", password = "postgres")

The next step typically involves checking if PostGIS was installed on the working database, and if not try to install it:

pgPostGIS(con)

The function should return TRUE for all pg- functions to work.

Finally, at the end of an interactive session, the connection to the database should be closed:

dbDisconnect(con)

Author

Mathieu Basille (mathieu@basille.org) and David Bucklin (david.bucklin@gmail.com)